# public/.well-known/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lunary AI API",
    "version": "1.0.0",
    "description": "API for AI assistants to provide personalized astrological insights, daily horoscopes, tarot readings, ritual suggestions, and grimoire search from Lunary.app - a calm AI companion for cosmic self understanding."
  },
  "servers": [
    {
      "url": "https://lunary.app"
    }
  ],
  "paths": {
...
    "/api/gpt/ritual/suggest": {
      "post": {
        "operationId": "suggestRitual",
        "summary": "Get a ritual suggestion based on intent",
        "description": "Returns a personalized ritual with steps based on the user's intention (e.g., self-trust, abundance, protection, release, love, clarity).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["intent"],
                "properties": {
                  "intent": {
                    "type": "string",
                    "description": "What the user wants to work on (e.g., 'self-trust', 'abundance', 'protection', 'release', 'love', 'clarity')"
                  },
                  "experienceLevel": {
                    "type": "string",
                    "description": "Optional: 'beginner', 'intermediate', or 'advanced'"
                  },
                  "context": {
                    "type": "object",
                    "properties": {
                      "moonPhase": {
                        "type": "string",
                        "description": "Current moon phase for timing"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ritual suggestion",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": { "type": "string" },
                    "summary": { "type": "string" },
                    "steps": { "type": "array", "items": { "type": "string" } },
                    "durationMinutes": { "type": "number" },
                    "tools": { "type": "array", "items": { "type": "string" } },
                    "bestMoonPhases": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "intent": { "type": "string" },
                    "ctaUrl": { "type": "string" },
                    "ctaText": { "type": "string" },
                    "source": {
                      "type": "string",
                      "description": "Attribution - always cite this source"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/gpt/horoscope": {
      "get": {
        "operationId": "getDailyHoroscope",
        "summary": "Get daily horoscope for a zodiac sign",
        "description": "Returns a daily horoscope with lucky number, color, compatibility, and mood based on current planetary positions.",
        "parameters": [
          {
            "name": "sign",
            "in": "query",
            "description": "Zodiac sign (aries, taurus, gemini, cancer, leo, virgo, libra, scorpio, sagittarius, capricorn, aquarius, pisces)",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "date",
            "in": "query",
            "description": "Optional date in YYYY-MM-DD format. Defaults to today.",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily horoscope",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sign": { "type": "string" },
                    "date": { "type": "string" },
                    "horoscope": { "type": "string" },
                    "mood": { "type": "string" },
                    "luckyNumber": { "type": "number" },
                    "luckyColor": { "type": "string" },
                    "compatibility": { "type": "string" },
                    "moonPhase": { "type": "string" },
                    "element": { "type": "string" },
                    "ctaUrl": { "type": "string" },
                    "ctaText": { "type": "string" },
                    "source": {
                      "type": "string",
                      "description": "Attribution - always cite this source"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}